refactor(api): split main.py into per-domain routers (#82)
Découpe `api/main.py` (628 lignes) en 5 routers FastAPI sous
`api/routers/`. Aucun changement de comportement HTTP : mêmes URLs,
payloads, media types, et structure OpenAPI.

Layout cible:
- `api/main.py` (165 lignes) : lifespan, app, `/`, `/health`, include_router(...)
- `api/routers/admin.py`       — 1 endpoint (api-keys)
- `api/routers/etl.py`         — 3 endpoints (run, jobs, job)
- `api/routers/flux.py`        — 7 endpoints + `_load_flux_df` helper +
                                  commentaire sur l'ordre des routes
                                  (catch-all `{table_name}` après `.xlsx`/`.arrow`)
- `api/routers/taxes.py`       — 6 endpoints (accise × 3, cta × 3)
- `api/routers/facturation.py` — 5 endpoints (rapport, detail × 2,
                                  documents.xlsx, check/odoo)

Conventions:
- Chaque router déclare `APIRouter(tags=["..."])` — tag dropped des ~12
  appels de décorateurs (FastAPI merge le tag du router au tag de l'endpoint)
- Pas de `prefix` URL : chaque endpoint porte son chemin complet
  (`@router.get("/taxes/accise/rapport.xlsx")`), greppable depuis l'URL
- `/` et `/health` restent dans `main.py` — couplés à l'état lifespan
  `_tg_app`, leur déplacement nécessiterait `app.state` (magie évitée)

Tests:
- ~30 monkey-patches `electricore.api.main.<fn>` → `electricore.api.routers.<domaine>.<fn>`
  dans 4 fichiers de tests (mécanique)
- Suite : 463 passed, 35 skipped, 0 failed

Closes #82
