+++ electricore/api/main.py
@@ -9,30 +9,15 @@ import os
 from contextlib import asynccontextmanager
 from urllib.parse import urlparse
 
-from fastapi import Depends, FastAPI, HTTPException, Query
+from fastapi import FastAPI
 
 from electricore.api.config import settings
-from electricore.api.decorators import arrow_endpoint, xlsx_endpoint
-from electricore.api.models import ETLJobResponse, ETLRunRequest
-from electricore.api.security import APIKeyInfo, get_api_key_info, get_current_api_key
-from electricore.api.serializers import arrow_stream, xlsx_multi_sheet
-from electricore.api.services import duckdb_service, etl_service
-from electricore.api.services.check_facturation_service import verifier_odoo
-from electricore.integrations.odoo.decorators import with_odoo
-from electricore.integrations.odoo.facturation import (
-    documents_facturation_du_mois,
-    facturation_du_mois,
-    feuilles_rapport_facturation,
-    rapport_facturation,
-)
-from electricore.integrations.odoo.taxes import (
-    accise_par_contrat,
-    cta_par_contrat,
-    feuilles_rapport_accise,
-    feuilles_rapport_cta,
-    rapport_accise,
-    rapport_cta,
-)
+from electricore.api.routers import admin as admin_router
+from electricore.api.routers import etl as etl_router
+from electricore.api.routers import facturation as facturation_router
+from electricore.api.routers import flux as flux_router
+from electricore.api.routers import taxes as taxes_router
+from electricore.api.services import duckdb_service
 
 logger = logging.getLogger(__name__)
 
@@ -109,6 +94,13 @@ app = FastAPI(
     ],
 )
 
+# Routers per-domaine (issue #82). Chaque router porte son tag OpenAPI et ses endpoints.
+app.include_router(admin_router.router)
+app.include_router(etl_router.router)
+app.include_router(flux_router.router)
+app.include_router(taxes_router.router)
+app.include_router(facturation_router.router)
+
 
 @app.get("/", tags=["public"])
 async def root():
@@ -144,140 +136,6 @@ async def root():
     }
 
 
-@xlsx_endpoint(app, "/flux/c15/entrees.xlsx", filename="entrees_c15.xlsx", error_status=500, tags=["flux"])
-def get_entrees_xlsx(
-    limit: int = Query(10000, le=100000, description="Nombre maximum de lignes"),
-) -> bytes:
-    """Exporte les **entrées** C15 au format XLSX (PMES, MES, CFNE)."""
-    from electricore.api.serializers import xlsx_multi_sheet
-    from electricore.core.loaders.duckdb import c15
-
-    df = c15().entrees().limit(limit).collect()
-    return xlsx_multi_sheet({"entrees": df})
-
-
-@xlsx_endpoint(app, "/flux/c15/sorties.xlsx", filename="sorties_c15.xlsx", error_status=500, tags=["flux"])
-def get_sorties_xlsx(
-    limit: int = Query(10000, le=100000, description="Nombre maximum de lignes"),
-) -> bytes:
-    """Exporte les **sorties** C15 au format XLSX (RES, CFNS)."""
-    from electricore.api.serializers import xlsx_multi_sheet
-    from electricore.core.loaders.duckdb import c15
-
-    df = c15().sorties().limit(limit).collect()
-    return xlsx_multi_sheet({"sorties": df})
-
-
-def _load_flux_df(table_name: str, prm: str | None, limit: int):
-    """Charge un flux via DuckDBQuery + serializer. Whitelist déclarative via FLUX_CONFIGS.
-
-    Test seam : monkeypatch ce helper pour court-circuiter l'IO DuckDB dans les
-    tests d'endpoint (le SQL est paramétré et la sécurité est testée séparément).
-    """
-    from electricore.core.loaders.duckdb.helpers import make_query
-    from electricore.core.loaders.duckdb.registry import FLUX_CONFIGS
-
-    if table_name not in FLUX_CONFIGS:
-        raise HTTPException(
-            404,
-            f"Table '{table_name}' non trouvée. Tables disponibles: {sorted(FLUX_CONFIGS.keys())}",
-        )
-
-    query = make_query(FLUX_CONFIGS[table_name])
-    if prm:
-        query = query.filter({"pdl": prm})
-    return query.limit(limit).collect()
-
-
-# IMPORTANT — ordre de déclaration : les routes `.xlsx` / `.arrow` et `/info`
-# (sous-ressource) DOIVENT précéder la route catch-all `/flux/{table_name}`,
-# sinon `{table_name}` capture `c15.xlsx` et la requête est mal routée.
-
-
-@xlsx_endpoint(app, "/flux/{table_name}.xlsx", filename="flux_{table_name}.xlsx", error_status=500, tags=["flux"])
-def get_flux_xlsx(
-    table_name: str,
-    prm: str | None = Query(None, description="Filtrer par pdl (Point de Livraison)"),
-    limit: int = Query(10000, le=100000, description="Nombre maximum de lignes"),
-) -> bytes:
-    """Exporte les données d'un flux Enedis au format XLSX (max 100 000 lignes)."""
-    from electricore.api.serializers import xlsx_multi_sheet
-
-    df = _load_flux_df(table_name, prm, limit)
-    return xlsx_multi_sheet({table_name: df})
-
-
-@arrow_endpoint(app, "/flux/{table_name}.arrow", tags=["flux"])
-def get_flux_arrow(
-    table_name: str,
-    prm: str | None = Query(None, description="Filtrer par pdl (Point de Livraison)"),
-    limit: int = Query(1_000_000, le=10_000_000, description="Nombre maximum de lignes (défaut 1 000 000)"),
-) -> bytes:
-    """Exporte les données d'un flux Enedis en flux Arrow IPC (max 10 000 000 lignes).
-
-    Consommable côté client avec `pl.read_ipc_stream(BytesIO(content))` — typage et
-    timezones préservés, pas de perte de précision contrairement à XLSX. Cf. `ElectricoreClient.flux`.
-    """
-    from electricore.api.serializers import arrow_stream
-
-    df = _load_flux_df(table_name, prm, limit)
-    return arrow_stream(df)
-
-
-@app.get("/flux/{table_name}/info", tags=["flux"])
-async def get_table_info(table_name: str, api_key: str = Depends(get_current_api_key)):
-    """
-    Retourne les métadonnées d'une table (colonnes, types, nombre de lignes).
-
-    **Authentification requise** - Utilisez votre clé API.
-
-    Utile pour comprendre la structure des données avant de faire des requêtes.
-
-    Args:
-        table_name: Nom de la table flux (r151, c15, r64, etc.)
-        api_key: Clé API (automatiquement validée)
-
-    Returns:
-        Dict avec les métadonnées de la table (colonnes, types, nombre de lignes)
-    """
-    try:
-        return duckdb_service.get_table_info(table_name)
-    except Exception:
-        available_tables = duckdb_service.list_tables()
-        raise HTTPException(404, f"Table '{table_name}' non trouvée. Tables disponibles: {available_tables}")
-
-
-# `/flux/{table_name}` (JSON) déclaré APRÈS les routes plus spécifiques pour que
-# `/flux/c15.xlsx` matche bien l'endpoint XLSX et pas le catch-all.
-@app.get("/flux/{table_name}", tags=["flux"])
-async def get_flux(
-    table_name: str,
-    prm: str | None = Query(None, description="Filtrer par pdl (Point de Livraison)"),
-    limit: int = Query(100, le=1000, description="Nombre maximum de lignes à retourner"),
-    offset: int = Query(0, ge=0, description="Nombre de lignes à ignorer (pagination)"),
-    api_key: str = Depends(get_current_api_key),
-):
-    """
-    Endpoint générique pour lire n'importe quel flux Enedis (réponse JSON).
-
-    **Authentification requise** - Utilisez votre clé API.
-
-    Exemples:
-    - /flux/r151 : Relevés quotidiens
-    - /flux/c15 : Changements contractuels
-    - /flux/r64 : Relevés demandés sur SGE
-    - /flux/f15 : Facturation Enedis détaillée
-    """
-    df = _load_flux_df(table_name, prm, limit + offset)
-    rows = df.slice(offset, limit).to_dicts()
-    return {
-        "table": f"flux_{table_name}",
-        "filters": {"pdl": prm} if prm else None,
-        "pagination": {"limit": limit, "offset": offset, "returned": len(rows)},
-        "data": rows,
-    }
-
-
 @app.get("/health", tags=["public"])
 async def health():
     """
@@ -305,324 +163,3 @@ async def health():
             "method": "X-API-Key header",
         },
     }
-
-
-@app.post("/etl/run", tags=["etl"], response_model=ETLJobResponse, status_code=202)
-async def run_etl(
-    body: ETLRunRequest,
-    api_key: str = Depends(get_current_api_key),
-):
-    """
-    Lance le pipeline d'ingestion ETL en arrière-plan.
-
-    **Authentification requise.**
-
-    Modes disponibles :
-    - `test` — 2 fichiers par flux (~3s), dataset `flux_enedis_test`
-    - `r151` — R151 complet, dataset `flux_enedis_r151`
-    - `all` — Tous les flux en production, dataset `flux_enedis`
-    - `reset` — Reset complet (supprime l'état incrémental), dataset `flux_enedis`
-
-    Retourne immédiatement un `job_id` pour suivre l'avancement via `GET /etl/jobs/{job_id}`.
-
-    Codes :
-    - 202 : job lancé
-    - 409 : un job est déjà en cours
-    - 501 : extra [etl] non installé
-    """
-    if not etl_service.is_etl_available():
-        raise HTTPException(501, "Le pipeline ETL n'est pas disponible. Installez l'extra [etl] : uv sync --extra etl")
-
-    try:
-        mode = etl_service.ETLMode(body.mode)
-    except ValueError:
-        raise HTTPException(422, f"Mode invalide : '{body.mode}'. Valeurs acceptées : test, r151, all, reset")
-
-    if etl_service.is_running():
-        raise HTTPException(409, "Un job ETL est déjà en cours d'exécution.")
-
-    job = await etl_service.start_job(mode)
-    return ETLJobResponse(
-        id=job.id,
-        mode=job.mode,
-        status=job.status,
-        started_at=job.started_at,
-        finished_at=job.finished_at,
-        error=job.error,
-        output=job.output,
-    )
-
-
-@app.get("/etl/jobs", tags=["etl"], response_model=list[ETLJobResponse])
-async def list_etl_jobs(
-    limit: int = Query(20, ge=1, le=50, description="Nombre de jobs à retourner"),
-    api_key: str = Depends(get_current_api_key),
-):
-    """
-    Liste les jobs ETL récents (ordre anti-chronologique).
-
-    **Authentification requise.**
-    """
-    jobs = etl_service.list_jobs(limit)
-    return [
-        ETLJobResponse(
-            id=j.id,
-            mode=j.mode,
-            status=j.status,
-            started_at=j.started_at,
-            finished_at=j.finished_at,
-            error=j.error,
-            output=j.output,
-        )
-        for j in jobs
-    ]
-
-
-@app.get("/etl/jobs/{job_id}", tags=["etl"], response_model=ETLJobResponse)
-async def get_etl_job(
-    job_id: str,
-    api_key: str = Depends(get_current_api_key),
-):
-    """
-    Retourne le statut d'un job ETL par son identifiant.
-
-    **Authentification requise.**
-
-    Statuts possibles : `running` | `completed` | `failed`
-    """
-    job = etl_service.get_job(job_id)
-    if job is None:
-        raise HTTPException(404, f"Job '{job_id}' introuvable.")
-    return ETLJobResponse(
-        id=job.id,
-        mode=job.mode,
-        status=job.status,
-        started_at=job.started_at,
-        finished_at=job.finished_at,
-        error=job.error,
-        output=job.output,
-    )
-
-
-@app.get("/admin/api-keys", tags=["admin"])
-async def list_api_keys(api_key: str = Depends(get_current_api_key), key_info: APIKeyInfo = Depends(get_api_key_info)):
-    """
-    Informations sur la configuration des clés API.
-
-    **Authentification requise** - Endpoint d'administration.
-
-    Returns:
-        Dict avec les informations sur les clés API configurées
-    """
-    return {
-        "message": "Configuration des clés API",
-        "current_key": {"preview": key_info.key_preview, "source": key_info.source},
-        "configuration": {
-            "total_keys": len(settings.get_valid_api_keys()),
-            "method": "X-API-Key header",
-            "public_endpoints": settings.public_endpoints,
-        },
-    }
-
-
-@xlsx_endpoint(
-    app, "/taxes/accise/rapport.xlsx", filename="accise_rapport{trimestre}.xlsx", requires_odoo=True, tags=["taxes"]
-)
-@with_odoo
-def export_accise_rapport_xlsx(
-    odoo,
-    trimestre: str | None = Query(
-        default=None,
-        examples=["2025-T1"],
-        description="Filtre par trimestre au format YYYY-TX. Sans filtre : toutes les données.",
-    ),
-) -> bytes:
-    """Livrable facturiste : Accise TICFE en XLSX multi-onglets (Résumé / Par taux / Détail)."""
-    return xlsx_multi_sheet(feuilles_rapport_accise(rapport_accise(odoo, trimestre)))
-
-
-@xlsx_endpoint(
-    app, "/taxes/accise/detail.xlsx", filename="accise_detail{trimestre}.xlsx", requires_odoo=True, tags=["taxes"]
-)
-@with_odoo
-def export_accise_detail_xlsx(
-    odoo,
-    trimestre: str | None = Query(
-        default=None,
-        examples=["2025-T1"],
-        description="Filtre par trimestre au format YYYY-TX. Sans filtre : toutes les données.",
-    ),
-) -> bytes:
-    """Détail Accise TICFE en XLSX mono-onglet (table PDL × mois, cas technique)."""
-    return xlsx_multi_sheet({"Détail": accise_par_contrat(odoo, trimestre)})
-
-
-@arrow_endpoint(app, "/taxes/accise/detail.arrow", requires_odoo=True, tags=["taxes"])
-@with_odoo
-def export_accise_detail_arrow(
-    odoo,
-    trimestre: str | None = Query(
-        default=None,
-        examples=["2025-T1"],
-        description="Filtre par trimestre au format YYYY-TX. Sans filtre : toutes les données.",
-    ),
-) -> bytes:
-    """Détail Accise TICFE en Arrow IPC stream (table PDL × mois, cas technique)."""
-    return arrow_stream(accise_par_contrat(odoo, trimestre))
-
-
-@xlsx_endpoint(
-    app, "/taxes/cta/rapport.xlsx", filename="cta_rapport{trimestre}.xlsx", requires_odoo=True, tags=["taxes"]
-)
-@with_odoo
-def export_cta_rapport_xlsx(
-    odoo,
-    trimestre: str | None = Query(
-        default=None,
-        examples=["2025-T1"],
-        description="Filtre par trimestre au format YYYY-TX. Sans filtre : toutes les données.",
-    ),
-) -> bytes:
-    """Livrable facturiste : CTA en XLSX multi-onglets (Résumé / Par taux / Détail)."""
-    return xlsx_multi_sheet(feuilles_rapport_cta(rapport_cta(odoo, trimestre)))
-
-
-@xlsx_endpoint(app, "/taxes/cta/detail.xlsx", filename="cta_detail{trimestre}.xlsx", requires_odoo=True, tags=["taxes"])
-@with_odoo
-def export_cta_detail_xlsx(
-    odoo,
-    trimestre: str | None = Query(
-        default=None,
-        examples=["2025-T1"],
-        description="Filtre par trimestre au format YYYY-TX. Sans filtre : toutes les données.",
-    ),
-) -> bytes:
-    """Détail CTA mensuel en XLSX mono-onglet (PDL × mois, cas technique)."""
-    return xlsx_multi_sheet({"Détail": cta_par_contrat(odoo, trimestre)})
-
-
-@arrow_endpoint(app, "/taxes/cta/detail.arrow", requires_odoo=True, tags=["taxes"])
-@with_odoo
-def export_cta_detail_arrow(
-    odoo,
-    trimestre: str | None = Query(
-        default=None,
-        examples=["2025-T1"],
-        description="Filtre par trimestre au format YYYY-TX. Sans filtre : toutes les données.",
-    ),
-) -> bytes:
-    """Détail CTA mensuel en Arrow IPC stream (PDL × mois, cas technique)."""
-    return arrow_stream(cta_par_contrat(odoo, trimestre))
-
-
-@xlsx_endpoint(
-    app,
-    "/facturation/rapport.xlsx",
-    filename="facturation_rapport{mois}.xlsx",
-    requires_odoo=True,
-    tags=["facturation"],
-)
-@with_odoo
-def export_facturation_rapport_xlsx(
-    odoo,
-    mois: str | None = Query(
-        default=None,
-        examples=["2025-01-01"],
-        description="Mois au format YYYY-MM-DD (défaut : dernier mois disponible dans les données)",
-    ),
-) -> bytes:
-    """Livrable facturiste : 3 onglets (Résumé / Lignes / Changements puissance)."""
-    return xlsx_multi_sheet(feuilles_rapport_facturation(rapport_facturation(odoo, mois)))
-
-
-@xlsx_endpoint(
-    app,
-    "/facturation/detail.xlsx",
-    filename="facturation_detail{mois}.xlsx",
-    requires_odoo=True,
-    tags=["facturation"],
-)
-@with_odoo
-def export_facturation_detail_xlsx(
-    odoo,
-    mois: str | None = Query(
-        default=None,
-        examples=["2025-01-01"],
-        description="Mois au format YYYY-MM-DD (défaut : dernier mois disponible dans les données)",
-    ),
-) -> bytes:
-    """Lignes brutes du rapprochement Odoo↔Enedis en XLSX mono-onglet (cas technique)."""
-    return xlsx_multi_sheet({"Détail": facturation_du_mois(odoo, mois)})
-
-
-@arrow_endpoint(app, "/facturation/detail.arrow", requires_odoo=True, tags=["facturation"])
-@with_odoo
-def export_facturation_detail_arrow(
-    odoo,
-    mois: str | None = Query(
-        default=None,
-        examples=["2025-01-01"],
-        description="Mois au format YYYY-MM-DD (défaut : dernier mois disponible dans les données)",
-    ),
-) -> bytes:
-    """Lignes brutes du rapprochement Odoo↔Enedis en Arrow IPC stream (cas technique)."""
-    return arrow_stream(facturation_du_mois(odoo, mois))
-
-
-@app.get("/facturation/check/odoo", tags=["facturation"])
-async def check_facturation_odoo(api_key: str = Depends(get_current_api_key)):
-    """
-    Vérifications pré-facturation côté Odoo.
-
-    **Authentification requise. Nécessite Odoo configuré.**
-
-    Pour tous les sale.order avec state='sale', retourne :
-    - `rsc_manquante` : liste des sale.order sans `x_ref_situation_contractuelle`
-    - `cfne_manquante` : liste des sale.order sans `x_date_cfne`
-    - `invoicing_state_counts` : répartition des `x_invoicing_state`
-    - `factures_draft` : factures encore en draft (anomalie après campagne)
-
-    Chaque entrée contient un lien direct vers l'enregistrement Odoo (champ `url`).
-    """
-    if not settings.is_odoo_configured:
-        raise HTTPException(
-            501,
-            f"Odoo [{settings.odoo_env}] non configuré. Définissez ODOO_{settings.odoo_env.upper()}_URL/DB/USERNAME/PASSWORD dans .env",
-        )
-    try:
-        result = await asyncio.get_event_loop().run_in_executor(None, verifier_odoo)
-    except Exception as e:
-        logger.exception("Erreur facturation/check/odoo")
-        raise HTTPException(503, f"Erreur lors de la vérification Odoo : {e}")
-    return result
-
-
-@xlsx_endpoint(
-    app,
-    "/facturation/documents.xlsx",
-    filename="facturation{mois}.xlsx",
-    requires_odoo=True,
-    tags=["facturation"],
-)
-@with_odoo
-def export_facturation_documents(
-    odoo,
-    mois: str | None = Query(
-        default=None,
-        examples=["2025-01-01"],
-        description="Mois au format YYYY-MM-DD (défaut : dernier mois disponible dans les données)",
-    ),
-) -> bytes:
-    """Livrable XLSX multi-onglets de tous les documents utiles pour la facturation (cf. #78).
-
-    Onglets :
-
-    - **F15 complet** : flux F15 du mois
-    - **F15 prestations** : F15 filtré sur les prestations (unite = 'UNITE')
-    - **C15 complet** : flux C15 du mois
-    - **C15 sorties** : C15 filtré sur les sorties (RES + CFNS)
-    - **Réconciliation** : réconciliation Odoo ↔ Enedis
-    - **Changements puissance** : lignes avec changement de puissance
-    """
-    documents, _suffix = documents_facturation_du_mois(odoo, mois)
-    return xlsx_multi_sheet(documents)
